+++ radio/src/tests/lua.cpp
@@ -42,10 +42,6 @@
 #define SWAP_DEFINED
 #include "opentx.h"
 
-#include <lua.h>
-#include <lauxlib.h>
-#include <lualib.h>
-
 extern const char * zchar2string(const char * zstring, int size);
 #define EXPECT_ZSTREQ(c_string, z_string)   EXPECT_STREQ(c_string, zchar2string(z_string, sizeof(z_string)))
 extern void luaInit();
